+++ src/Components/StateMachine/Subscriber/TransitionSubscriber.php
@@ -19,6 +19,7 @@ use Billie\BilliePayment\Util\CriteriaHelper;
 use Shopware\Core\Checkout\Order\Aggregate\OrderDelivery\OrderDeliveryCollection;
 use Shopware\Core\Checkout\Order\Aggregate\OrderDelivery\OrderDeliveryDefinition;
 use Shopware\Core\Checkout\Order\Aggregate\OrderDelivery\OrderDeliveryEntity;
+use Shopware\Core\Checkout\Order\Aggregate\OrderDelivery\OrderDeliveryStates;
 use Shopware\Core\Checkout\Order\OrderCollection;
 use Shopware\Core\Checkout\Order\OrderDefinition;
 use Shopware\Core\Checkout\Order\OrderEntity;
@@ -73,13 +74,13 @@ class TransitionSubscriber implements EventSubscriberInterface
         }
 
         switch ($event->getToPlace()->getTechnicalName()) {
-            case $this->configService->getStateForShip():
+            case OrderDeliveryStates::STATE_SHIPPED:
                 $this->operationService->ship($order);
                 break;
-            case $this->configService->getStateCancel():
+            case OrderDeliveryStates::STATE_CANCELLED:
                 $this->operationService->cancel($order);
                 break;
-            case $this->configService->getStateReturn():
+            case OrderDeliveryStates::STATE_RETURNED:
                 $this->operationService->return($order);
                 break;
         }